hexgrid: Switch memory layout for hexagonal hexgrid
The tiles of a hexagonal grid is now laid out in memory as 4 regions. 1
single cell region for the centre tile, and 3 rhombus shaped regions
that covers the remaining cells.

Which region a position belongs can determined by the sign of the
components and the offset within the region is given in a closed form
after that.

This leads to simpler code and a slight performance increase at the cost
of less natural iteration order.

Fixes #119